Add initial File persistence Get
