--- src/BapesBot.Service/Commands/Settings/CommandSettings.cs
@@ -1,10 +1,0 @@
-using Newtonsoft.Json;
-
-namespace BapesBot.Service.Commands.Settings
-{
-    public class CommandSettings
-    {
-        [JsonIgnore]
-        public Command Command { get; set; }
-    }
-}
