--- src/BapesBot.Service/Commands/Settings/CounterSettings.cs
@@ -1,9 +1,0 @@
-namespace BapesBot.Service.Commands.Settings
-{
-    public class CounterSettings : CommandSettings
-    {
-        public string Key { get; set; }
-
-        public int Value { get; set; }
-    }
-}
