+++ src/GolfClapBot.Service/Persistence/ICommandPersistence.cs
@@ -1,9 +1,11 @@
+using GolfClapBot.Service.Commands;
+
 namespace GolfClapBot.Service.Persistence
 {
     public interface ICommandPersistence<T> where T : CommandSettings.CommandSettings
     {
         void Save(T value);
 
-        T Get();
+        T Get(Command command);
     }
 }
